requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1

# Ingestion
PyMuPDF==1.25.1
pdfplumber==0.11.4
Pillow==11.0.0

# Embeddings
FlagEmbedding==1.2.11
open-clip-torch==2.26.1
torch==2.5.1
torchvision==0.20.1
transformers==4.46.3
sentence-transformers==3.3.1

# Vector store
faiss-cpu==1.9.0.post1
numpy==1.26.4

# RAG / LLM
langchain==0.3.7
langchain-community==0.3.7
langchain-openai==0.2.9
openai==1.54.4

# API + UI
fastapi==0.115.5
uvicorn[standard]==0.32.1
gradio==5.6.0
python-multipart==0.0.17

# Evaluation
ragas==0.2.6
datasets==3.1.0

# Dev / Testing
pytest==8.3.3
httpx==0.27.2
